requirements.txt
requests==2.32.3
PyYAML==6.0.2
schedule==1.2.2
loguru==0.7.2
tenacity==9.0.0
pytest==8.3.5
